MLops - ML + Flask + Celery/requirements.txt
torch==1.5.1
numpy==1.18.4
Flask==1.1.1
requests==2.22.0
torchvision==0.6.0
Pillow==7.2.0
